Refactoring FIT's and fixing a new failure.
